### 2018-03-06 #603 @whedon
[PRE REVIEW]: microPIECE (microRNA pipeline enhanced by CLIP experiments)
**Submitting author:** @greatfireball (<a href="http://orcid.org/0000-0003-4166-5423">Frank Förster</a>)
**Repository:** <a href="https://github.com/microPIECE-team/microPIECE" target ="_blank">https://github.com/microPIECE-team/microPIECE</a>
**Version:** v1.1.0
