### 2018-03-06 65976bb
Bump to version v1.0.4
