### 2018-03-06 #603 @DanielAmsel
Sorry to see that the PDF compilation step failed.
We tested this with our docker container and it worked:
`greatfireball/docker_whedon_translate4joss`

Any ideas how to fix that?
