### 2018-03-06 d55873b
Using docker container for testing of content in scripts folder
