+++ .travis.yml
@@ -17,7 +17,7 @@ before_install:
 install:
     - cpanm Devel::Cover Devel::Cover::Report::Coveralls Test::Script::Run Log::Log4perl
 script:
-    - prove -It t/*.t
+    - docker run -it --rm -v $PWD:/opt/microPIECE -v /tmp/microPIECE-testset:/tmp/microPIECE-testset micropiece/micropiece /bin/bash -c 'cd /opt/microPIECE; PERL5OPT="-MDevel::Cover=-coverage,statement,branch,condition,path,subroutine,time" prove -It t/*.t'
     - docker run -it --rm -v $PWD:/opt/microPIECE -v /tmp/microPIECE-testset:/tmp/microPIECE-testset micropiece/micropiece /opt/microPIECE/t/run_clip.sh
     - docker run -it --rm -v $PWD:/opt/microPIECE -v /tmp/microPIECE-testset:/tmp/microPIECE-testset micropiece/micropiece /opt/microPIECE/t/run_complete.sh
     - docker run -it --rm -v $PWD:/opt/microPIECE -v /tmp/microPIECE-testset:/tmp/microPIECE-testset micropiece/micropiece chmod -R a+wrX /opt/microPIECE/cover_db
