### 2018-03-07 df14af7
removed 's' from parameter --mirna in help
