### 2018-03-07 bd72c5c
Bump version to v1.0.5
