+++ microPIECE.pl
@@ -275,6 +275,10 @@ Please report any new issues ad L<new Github-Issue|https://github.com/microPIECE
 
 =over 4
 
+=item v1.0.5 (2018-03-06)
+
+Update of documentation and correct spelling of C<--mirna> parameter
+
 =item v1.0.4 (2018-03-06)
 
 Fixes complete mature in final output (Fixes #69)
