### 2018-03-07 fa6f1e8
Refactored scripts/MINING_ortholog_blast.pl
