+++ microPIECE.pl
@@ -259,6 +259,10 @@ precursor microRNA set, containing novels (if mined), together with the known mi
 
 Semicolon-separated file --> rpm;condition;miRNA
 
+=item orthologous prediction file: F<miRNA_orthologs.csv>
+
+tab-separated file --> C<query_id subject_id identity aln_length num_mismatches num_gapopen query_start query_end subject_start subject_end evalue bitscore query_aligned_seq subject_aligned_seq query_length subject_length query_coverage subject_coverage>
+
 =item miRDeep2 mining result in HTML/CSV F<mirdeep_output.html/csv>
 
 the standard output HTML/CSV file of miRDeep2
