### 2018-03-07 de47555
Fix deletion of gap signs
