### 2018-03-07 def58e6
Add parameter --tempdir
The parameter --tempdir specifies where all temporary files are created.

Default location is the folder tmp inside the output directory,
therefore the default is

	out/tmp/

Fix #66
Fix #73
