### 2018-03-08 44f4dfd
Bump version to v1.0.6
