### 2018-03-08 2680690
updated changlog to v1.0.7
added bin size hot fix message
