### 2018-03-08 e3f63ec
Added a --piranahbinsize parameter
Fixes #80
