### 2018-03-08 4c2e9d1
Fix documentation for --piranahbinsize parameter
