### 2018-03-08 9bbb394
Add old 072_create_mirbase_struct.pl
