+++ scripts/072_create_mirbase_struct.pl
@@ -1,0 +1,163 @@
+#! /usr/bin/perl
+use strict;
+use warnings;
+use RNA::HairpinFigure qw/draw/;
+use Getopt::Long;
+use File::Temp qw(tmpnam);
+
+#>cel-let-7 (-42.90)   [cel-let-7-5p:17-38] [cel-let-7-3p:60-81]
+#
+#------uaca    gga             U              ---  aaua 
+#          cugu   uccggUGAGGUAG AGGUUGUAUAGUUu   gg    u
+#          ||||   ||||||||||||| ||||||||||||||   ||     
+#          gaca   aggCCAUUCCAUC UUUAACGUAUCaag   cc    u
+#agcuucucaa    --g             U              ugg  acca 
+
+
+#my $name   = 'hsa-mir-92a-1 MI0000093 Homo sapiens miR-92a-1 stem-loop';
+#my $seq    = 'CUUUCUACACAGGUUGGGAUCGGUUGCAAUGCUGUGUUUCUGUAUGGUAUUGCACUUGUCCCGGCCUGUUGAGUUUGG';
+#my $struct = '..(((...((((((((((((.(((.(((((((((((......)))))))))))))).)))))))))))).))).....';
+#my $figure = draw( $seq, $struct );
+#print ">$name\n$seq\n$struct\n$figure\n";
+
+
+#>NC_007417.3|344915..344975|-|126.9|result_bwt1.csvresult_bwa.csv|ame-miR-6000a-3p|high_conf
+#GGUUGGCAUAAGGUGGUACCAUGUAACAUUUUAACCCAUAGUACGACCCAUGCCGACUCA
+#(((((((((..(((.((((.(((.............))).)))).))).))))))))).. (-21.92)
+
+my $hairpin_file;
+my $mature_file;
+my $struct_file;
+
+GetOptions(
+	"hairpin=s"	=> \$hairpin_file,
+	"mature=s"	=> \$mature_file,
+	"struct=s"	=> \$struct_file) || die;
+
+my $rnafold		= "RNAfold -noPS";
+
+
+my %hairpin_hash	= %{&read_fasta($hairpin_file)};
+my %mature_hash		= %{&read_fasta($mature_file)};
+
+my %structure_hash	= %{&parse_structure($struct_file)};
+
+my $custom_tmp	= "custom.str";
+open(OUT,">",$custom_tmp) || die;
+
+foreach(keys %hairpin_hash){
+	my $hairpin_id		= $_;
+	my $hairpin_seq		= $hairpin_hash{$hairpin_id};
+	if(not exists $mature_hash{"$hairpin_id-5p"}){
+		print OUT "$structure_hash{$hairpin_id}";
+		print STDERR "Missing one arm of $hairpin_id : Using miRBase entry instead.\n";
+	}
+	elsif(not exists $mature_hash{"$hairpin_id-3p"}){
+		print OUT "$structure_hash{$hairpin_id}";
+		print STDERR "Missing one arm of $hairpin_id : Using miRBase entry instead.\n";
+	}
+	else{
+		my $mature5p_id		= "$hairpin_id-5p";
+		$mature5p_id		=~s/>//;
+		my $mature3p_id		= "$hairpin_id-3p";
+		$mature3p_id		=~s/>//;
+		my $mature5p_seq	= $mature_hash{">$mature5p_id"};
+		my $mature3p_seq	= $mature_hash{">$mature3p_id"};
+		my $mature5p_start	= index($hairpin_seq,$mature5p_seq)+1;
+		my $mature5p_stop	= $mature5p_start + length($mature5p_seq)-1;
+		my $mature3p_start	= index($hairpin_seq,$mature3p_seq)+1;
[…]
