### 2018-03-08 1821e46
Add new parameter --out and improved error handling
