### 2018-03-09 #603 @arfon
@brainstorm @mgymrek - would either of you be willing to edit this submission for JOSS?
