### 2018-03-09 8647ff2
Set required base docker image version to 1.2
