### 2018-03-10 53ce065
Refactoring of ISOMIR_reformat_isomirs.pl
