+++ scripts/ISOMIR_reformat_isomirs.pl
@@ -8,79 +8,83 @@ my $condition_ID	= $ARGV[1];	# from db
 
 my @input_array		= split(",",$input_list);
 my %condition_hash;	#{condition_file}	= \%rep_hash;
+
+my @required_fields = qw(seq freq mir mism add t5 t3 ambiguity);
+
+my %rep_hash;	 #{miRBaseID;mism;add;t5;t3;ambiguity} = freq/ambiguous
+
 foreach my $input_file	(@input_array){
-	my %rep_hash;	 #{miRBaseID;mism;add;t5;t3;ambiguity} = freq/ambiguous
-	open(TMP,"<",$input_file) || die;
-	my $tmp_firstline	= <TMP>;
-	my $tmp_total_read_count= 0;
-	while(<TMP>){
-		chomp;
-		my @tmp_split	= split("\t",$_);
-		my $tmp_seq	= $tmp_split[0];
-		my $tmp_freq	= $tmp_split[2];
-		my $tmp_id	= $tmp_split[3];
-		my $tmp_mism	= $tmp_split[6];
-		my $tmp_add	= $tmp_split[7];
-		my $tmp_t5	= $tmp_split[8];
-		my $tmp_t3	= $tmp_split[9];
-		my $tmp_amb	= $tmp_split[14];
-		my $tmp_div_c	= $tmp_freq/$tmp_amb;
+    open(TMP,"<",$input_file) || die "Unable to open input file '$input_file': $!\n";
+    
+    my $fieldlist	= <TMP>;
+    chomp($fieldlist);
+    my @possible_fields = split(/\s+/, $fieldlist);
+
+    # check if required fields are existing
+    foreach my $req_field (@required_fields)
+    {
+	unless (grep {$_ eq $req_field} (@possible_fields))
+	{
+	    die("Missing field '$req_field' in input file '$input_file'\n");
+	}
+    }
+    
+    my $tmp_total_read_count= 0;
+    while(<TMP>)
+    {
+	chomp;
+	my %fields = ();
+	@fields{@possible_fields} = split("\t", $_);
+
+	# check if one of the mism add t5 t3 fields is 0
+	foreach my $field (qw(mism add t5 t3))
+	{
+	    if ($fields{$field} eq "0")
+	    {
+		$fields{$field} = 'NULL';
+	    }
+	}
+
+	my $div_c = $fields{freq}/$fields{ambiguity};
 	
-		if($tmp_mism eq 0){
-			$tmp_mism = 'NULL';
-		}
-		if($tmp_add eq 0){
-			$tmp_add = 'NULL';
-		}
-		if($tmp_t5 eq 0){
-			$tmp_t5 = 'NULL';
-		}
-		if($tmp_t3 eq 0){
-			$tmp_t3 = 'NULL';
-		}
-		my $tmp_key	= "$tmp_id;$tmp_mism;$tmp_add;$tmp_t5;$tmp_t3;$tmp_seq";
-		$tmp_total_read_count += ($tmp_freq/$tmp_amb);	# count all reads, divided by ambiguity, but also the non-isoforms
-		next if(($tmp_mism eq 'NULL') && ($tmp_add eq 'NULL') && ($tmp_t5 eq 'NULL') && ($tmp_t3 eq 'NULL'));
-		$rep_hash{$tmp_key}= $tmp_div_c;
-	#	print ">>> $tmp_key\n";
-	}
-	close(TMP);
-	# calc RPM
-	my %rpm_hash;
-	foreach my $rep_key (keys %rep_hash){
-		my $rep_freq 	= $rep_hash{$rep_key};
[…]
