### 2018-03-11 90a60d4
Add infos about isomir output in files
