### 2018-03-11 f830bde
Add description about genomic location output file
