### 2018-03-12 e91bcb7
Bump version to v1.1.0
