### 2018-03-12 #603 @brainstorm
Sure! Happy to :)

Den 10 mars 2018 8:34 fm skrev "Arfon Smith" <notifications@github.com>:

> @brainstorm <https://github.com/brainstorm> @mgymrek
> <https://github.com/mgymrek> - would either of you be willing to edit
> this submission for JOSS?
>
