### 2018-03-12 #603 @arfon
> Sure! Happy to :)

Great! Next step is to look for a reviewer, probably someone from here: http://bit.ly/joss-reviewers
