### 2018-03-12 d3f88b2
Add information about docker images into README.md
