### 2018-03-12 f436436
Update --piranahbinsize parameter description
