### 2018-03-12 eef8a0e
Add enumeration for miRNA expression output
