### 2018-03-12 21225d3
Add enumeration for miRNA orthologues sequences output
