### 2018-03-12 4b0cc54
Add enumeration for isomir detection result
