+++ microPIECE.pl
@@ -415,18 +415,18 @@ Please report any new issues ad L<new Github-Issue|https://github.com/microPIECE
 
 =over 4
 
-=item v1.1.0 (2018-03-12)
+=item L<v1.1.0|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.1.0> (2018-03-12)
 
 Add isomir detection and copy the final genomic location file to the
 output filter (Fixes
 L<#34|https://github.com/microPIECE-team/microPIECE/issues/34>)
 
-=item v1.0.7 (2018-03-08)
+=item L<v1.0.7|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.7> (2018-03-08)
 
 Piranha was lacking of a bin_size parameter. Added parameter C<--piranahbinsize> with a default value of C<20>
 (Fixes L<#66|https://github.com/microPIECE-team/microPIECE/issues/80>)
 
-=item v1.0.6 (2018-03-08)
+=item L<v1.0.6|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.6> (2018-03-08)
 
 Added parameter C<--mirbasedir> and C<--tempdir> to support local
 mirbase files and relocation of directory for temporary files (Fixes
@@ -434,27 +434,27 @@ L<#66|https://github.com/microPIECE-team/microPIECE/issues/66>,
 L<#73|https://github.com/microPIECE-team/microPIECE/issues/73>, and
 L<#76|https://github.com/microPIECE-team/microPIECE/issues/76>)
 
-=item v1.0.5 (2018-03-07)
+=item L<v1.0.5|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.5> (2018-03-07)
 
 Update of documentation and correct spelling of C<--mirna> parameter
 
-=item v1.0.4 (2018-03-07)
+=item L<v1.0.4|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.4> (2018-03-07)
 
 Fixes complete mature in final output (Fixes L<#69|https://github.com/microPIECE-team/microPIECE/issues/69>)
 
-=item v1.0.3 (2018-03-06)
+=item L<v1.0.3|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.3> (2018-03-06)
 
 Add tests for perl scripts in script folder which ensure the correct handling of BED stop coordinates (Fixes L<#65|https://github.com/microPIECE-team/microPIECE/issues/65>)
 
-=item v1.0.2 (2018-03-05)
+=item L<v1.0.2|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.2> (2018-03-05)
 
 Fixes the incorrect sorting of BED files, result was correct, but sorting was performed in the wrong order. (Fixes L<#63|https://github.com/microPIECE-team/microPIECE/issues/63>)
 
-=item v1.0.1 (2018-03-05)
+=item L<v1.0.1|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.1> (2018-03-05)
 
 Fix an error conserning BED file handling of start and stop coordinates. (Fixes L<#59|https://github.com/microPIECE-team/microPIECE/issues/59>)
 
-=item v1.0.0 (2018-03-05)
+=item L<v1.0.0|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.0> (2018-03-05)
 
 =begin html
 
@@ -469,7 +469,7 @@ is archived as L<https://doi.org/10.5281/zenodo.1188484> and submitted to L<The 
 
 =end text
 
-=item v0.9.0 (2018-03-05)
+=item L<v0.9.0|https://github.com/microPIECE-team/microPIECE/releases/tag/v0.9.0> (2018-03-05)
 
 =begin html
 
