+++ pull_request_template.md
@@ -1,0 +1,8 @@
+Fixes # .
+
+Changes proposed within this pull request
+-
+-
+-
+
+@microPIECE-team/micropiece-team-admins
