### 2018-03-13 2fc9137
Create CONTRIBUTING.md
Fixes #90
