+++ CONTRIBUTING.md
@@ -1,0 +1,31 @@
+# Introduction
+
+Thank you for considering contributing to the microPIECE software.
+
+Following these guidelines helps the developers of this open source project.
+Making it easier for them to address your issue, assess changes, and help you finalize your pull requests.
+Any kind of contribution, to code, documentation, examples or tests are welcome.
+
+# Ground Rules
+Please have a look at our [code of conduct](CODE_OF_CONDUCT.md).
+
+# Contributing Code
+Please follow this procedure
+1. Open an issue with a bug report or feature request (indicate that you are willing to do the coding yourself)
+2. Create your own fork of the code
+3. Do the changes in your fork
+4. Open a pull request
+
+# How to report a bug
+
+When filing an issue, make sure to answer those questions:
+
+1. What version of Perl and chloroExtractor are you using?
+2. What operating system are you using?
+3. What did you do?
+4. What did you expect to see?
+5. What did you see instead?
+
+# How to suggest a feature or enhancement
+
+Open an issue which describes the feature you would like to see, why you need it, and how it should work.
