### 2018-03-13 fa3e3fe
changed pipeline name to microPIECE
