### 2018-03-15 2f3b820
First version our own binning tool for Piranha
