### 2018-03-15 6a61920
Generating Bins for a given BAM
