### 2018-03-15 257f670
Version with fast forward option
