### 2018-03-15 e745d75
Intoducing a progress bar
