### 2018-03-15 de4d1e2
Using Term::ProgressBar for the progressbar
