### 2018-03-15 489ab58
Avoid sorting to have the same sorting as the BAM has
