### 2018-03-15 bb79cec
First version for transcript-bed generation script
