+++ scripts/CLIP_generate_transcript_bed_sorted_by_bed.pl
@@ -1,0 +1,6 @@
+#!/usr/bin/env perl
+
+use strict;
+use warnings;
+
+use Getopt::Long;
