### 2018-03-15 352d8d5
Reading the parameter and the sort order file
