### 2018-03-15 c25d559
Add a output of the identified sort order
