### 2018-03-15 7700062
Importing GFF, skipping everything, but the requested feature and sorting according to the sortby file
