### 2018-03-15 787757c
sort order file might contain only sort order
