### 2018-03-15 12c9221
Warning and skipping non existent chromosomes
