### 2018-03-19 03d68f3
New script to create binned beds for Prianha
