+++ scripts/CLIP_binned_bed_from_bam_and_transcripts_for_piranha.pl
@@ -1,0 +1,238 @@
+#!/usr/bin/env perl
+
+use strict;
+use warnings;
+
+use Getopt::Long;
+
+my @bamfiles     = ();
+my @transcripts  = ();
+my $binsize      = 20;
+my $reqfeature   = "exon";
+my $pseudocounts = 0;  
+
+GetOptions(
+    'b|bam=s@'         => \@bamfiles,
+    's|size=i'         => \$binsize,
+    't|transcripts=s@' => \@transcripts,
+    'r|reqfeature=s'   => \$reqfeature,
+    ) || die;
+
+# definde the bams and transcripts if comma seperated
+@bamfiles = split(",", join(",", @bamfiles));
+@transcripts = split(",", join(",", @transcripts));
+
+# try to call samtools
+my $samtools = qx(samtools --version);
+if ($? != 0)
+{
+    die "Wrong version or no samtools available\n";
+}
+
+use Term::ProgressBar;
+
+# import the transcripts
+my %imported_transcripts = ();
+foreach my $transcript (@transcripts)
+{
+    open(FH, "<", $transcript) || die "Unable to open transcript file '$transcript'\n";
+    while (<FH>)
+    {
+	chomp($_);
+
+	# as the input is a gff file, we want to consider only the required feature
+	my ($chr, undef, $feature, $start, $end, undef, $strand) = split(/\t/, $_);
+	next unless ($feature eq $reqfeature);
+
+	# transform the start and end coordinates into BED format (0-based, end-exclusive)
+	($start, $end) = transform2bed($start, $end);
+
+	# strand + => 0; - => 1;
+	if ($strand eq "+")
+	{
+	    $strand = 0;
+	} elsif ($strand eq "-")
+	{
+	    $strand = 1;
+	}
+
+	# start and stop are transformed into bins
+	$start = int($start/$binsize);
+	$end   = int(($end-1)/$binsize);
+	push(@{$imported_transcripts{$chr}{$strand}}, { start => $start, stop => $end });
+
+	$pseudocounts = 1;
+    }
+    close(FH) || die "Unable to close transcript file '$transcript'\n";
+}
+# finally sort the imported transcripts by coordinate
+foreach my $chr (keys %imported_transcripts)
+{
+    foreach my $strand (keys %{$imported_transcripts{$chr}})
+    {
+	my @bins_with_feature = ();
+	
+	foreach my $exon (@{$imported_transcripts{$chr}{$strand}})
+	{
+	    foreach my $bin ($exon->{start}..$exon->{stop})
+	    {
+		$bins_with_feature[$bin]++;
+	    }
+	}
+	$imported_transcripts{$chr}{$strand} = \@bins_with_feature;
+    }
+}
+
+foreach my $bam (@bamfiles)
+{
+    my $cmd = "samtools view -H $bam";
+    my $header = qx($cmd);
+    if ($? != 0)
+    {
+	die "Error running command '$cmd'\n";
+    }
+    my $seq = parse_header(\$header);
+    my $total_length = 0;
+    foreach my $seq (@{$seq})
+    {
+	$total_length += $seq->{len};
+    }
+
+    foreach my $seq (@{$seq})
+    {
+	my $progress = Term::ProgressBar->new({name => $seq->{name}, count => $seq->{len}, remove => 0});
+	$progress->minor(0);
+	my $next_val = 0;
+
+	my @counts = ();
+	$#counts = int($seq->{len}/$binsize)+1;
+
[…]
