### 2018-03-19 03430b5
Fixed samtools call
