### 2018-03-21 c749c77
Reduced memory footprint
