### 2018-03-21 3c136ed
Further reduced memory footprint
