### 2018-03-21 c990783
Add another progress bar and optimized expand_transcripts subroutine
