### 2018-03-22 9b7159e
Add better error description during file opening
