### 2018-03-22 110baf5
Add better error description during file closing
