### 2018-03-22 37659db
Output content better to read
