### 2018-03-22 0b5b51f
Refactor of the mirdeep parser
