+++ scripts/MINING_curate_mirdeep2fasta.pl
@@ -24,47 +24,46 @@ die "Need to specify --hairpinout file\n" unless (defined $hairpin_file);
 die "Need to specify --species three-letter-species-code\n" unless (defined $species);
 
 # get novel miRNAs above threshold
-my %novel_hash	= %{&parse_mirdeep($csv_file,$cutoff)};
+my $novels = parse_mirdeep($csv_file, $cutoff);
 
 open(MATURE,">",$mature_file)   || die "Unable to open file '$mature_file': $!\n";
 open(HAIRPIN,">",$hairpin_file) || die "Unable to open file '$hairpin_file': $!\n";
 
-foreach(keys %novel_hash){
-	my $novel_count	= $_;
-	my @novel_split	= @{$novel_hash{$novel_count}};
-	my $mature	= uc($novel_split[7]);
-	my $star	= uc($novel_split[8]);
-	my $hairpin	= uc($novel_split[9]);
-	
-	my $mature5p;
-	my $mature3p;
-	
-	my $mature_idx	= index($hairpin,$mature);
-	my $star_idx	= index($hairpin,$star);
-	
-	if($mature_idx < $star_idx){
-		$mature5p=$mature;
-		$mature3p=$star;
-	}
-	elsif($star_idx < $mature_idx){
-		$mature5p=$star;
-		$mature3p=$mature;
-	}
-	else{
-		die("mature and star sequence have the same position on hairpin\n"); # should never happen
-	}	
+for(my $novel_count = 1; $novel_count <= @{$novels}; $novel_count++)
+{
+    my $novel = $novels->[$novel_count-1];
+    
+    my $mature	= uc($novel->{mature_seq});
+    my $star	= uc($novel->{star_seq});
+    my $hairpin	= uc($novel->{precursor_seq});
 
-	$mature5p	=~ s/U/T/g;
-	$mature3p	=~ s/U/T/g;
-	$hairpin	=~ s/U/T/g;
+    my $mature5p;
+    my $mature3p;
 
-	my $header	= sprintf(">%s-new-%d", $species, $novel_count);
+    my $mature_idx	= index($hairpin,$mature);
+    my $star_idx	= index($hairpin,$star);
 
-        print HAIRPIN $header, "\n", $hairpin, "\n";
-        print MATURE  $header, "-5p\n", $mature5p, "\n", $header, "-3p\n", $mature3p, "\n";
+    if($mature_idx < $star_idx){
+	$mature5p=$mature;
+	$mature3p=$star;
+    }
+    elsif($star_idx < $mature_idx){
+	$mature5p=$star;
+	$mature3p=$mature;
+    }
+    else{
+	die("mature and star sequence have the same position on hairpin\n"); # should never happen
+    }
+
+    $mature5p	=~ s/U/T/g;
+    $mature3p	=~ s/U/T/g;
+    $hairpin	=~ s/U/T/g;
+
+    my $header	= sprintf(">%s-new-%d", $species, $novel_count);
+
+    print HAIRPIN $header, "\n", $hairpin, "\n";
+    print MATURE  $header, "-5p\n", $mature5p, "\n", $header, "-3p\n", $mature3p, "\n";
 }
-#close(CSV) || die;
-
 close(MATURE) || die "Unable to close file '$mature_file': $!\n";
 close(HAIRPIN)|| die "Unable to close file '$hairpin_file': $!\n";
 
@@ -72,33 +71,56 @@ close(HAIRPIN)|| die "Unable to close file '$hairpin_file': $!\n";
 
 # skip all parts but novel section
 sub parse_mirdeep{
-        my $pm_file     = $_[0];
-        my $pm_cutoff   = $_[1];
-        my %pm_hash;
[…]
