### 2018-03-22 87227a6
Using a checksum for number estimation of new miRNAs
