### 2018-03-23 e0aca91
Implemented a test for collitions
