### 2018-03-23 31d3eb3
Implement handling for non-existing input file
