### 2018-03-23 372cb0b
Implement test for existing hairpin file
