### 2018-03-23 e8362bb
Bump to version v1.2.2 and added descriptions for v1.2.0/v1.2.1
