### 2018-03-23 72ad725
Set microbadger links to new release v1.2.2
