+++ README.md
@@ -38,7 +38,7 @@ We also provide `microPIECE` as [DOCKER image](https://hub.docker.com/r/micropie
 ### Information about the docker images:
 | Branch | Size | Layers | Comment |
 |-|-|-|-|
-|[![](https://images.microbadger.com/badges/version/micropiece/micropiece:v1.1.0.svg)](https://microbadger.com/images/micropiece/micropiece:v1.1.0) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:v1.1.0.svg)](https://microbadger.com/images/micropiece/micropiece:v1.1.0) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:v1.1.0.svg)](https://microbadger.com/images/micropiece/micropiece:v1.1.0) | Latest release [v1.1.0](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.1.0) |
+|[![](https://images.microbadger.com/badges/version/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | Latest release [v1.2.2](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.2) |
 |[![](https://images.microbadger.com/badges/version/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | |
 |[![](https://images.microbadger.com/badges/version/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | |
 
