### 2018-03-23 bd28100
[CLIP binning]Add a warning, if multiple bams are provided
