### 2018-03-23 179e750
Correct md5sums for bins with pseudocounts
