### 2018-03-25 45ccae9
account for single matures from mirbase
that have a -[3|5]p extension.
