### 2018-03-26 8e7c153
Add more specific regular expression for converting mature into precursor name
