### 2018-03-26 6c6550a
Bump to version v1.2.3
