+++ README.md
@@ -38,7 +38,7 @@ We also provide `microPIECE` as [DOCKER image](https://hub.docker.com/r/micropie
 ### Information about the docker images:
 | Branch | Size | Layers | Comment |
 |-|-|-|-|
-|[![](https://images.microbadger.com/badges/version/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:v1.2.2.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.2) | Latest release [v1.2.2](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.2) |
+|[![](https://images.microbadger.com/badges/version/micropiece/micropiece:v1.2.3.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.3) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:v1.2.3.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.3) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:v1.2.3.svg)](https://microbadger.com/images/micropiece/micropiece:v1.2.3) | Latest release [v1.2.3](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.3) |
 |[![](https://images.microbadger.com/badges/version/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:master.svg)](https://microbadger.com/images/micropiece/micropiece:master) | |
 |[![](https://images.microbadger.com/badges/version/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | [![](https://images.microbadger.com/badges/image/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | [![](https://images.microbadger.com/badges/commit/micropiece/micropiece:develop.svg)](https://microbadger.com/images/micropiece/micropiece:develop) | |
 
@@ -281,11 +281,15 @@ Complete list of open issues is available on [Github-Issues](https://github.com/
 Please report any new issues ad [new Github-Issue](https://github.com/microPIECE-team/microPIECE/issues/new).
 
 ## Changelog
+- [v1.2.3](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.3) (2018-03-26)
+
+    Fix transformation of precursor sequences based on mirbase #22 precursor sequences with a single mature.
+    (Fixes L<#109|https://github.com/microPIECE-team/microPIECE/issues/109>)
 
 - [v1.2.2](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.2) (2018-03-23)
 
     Improved collision detection for newly identified miRNAs avoiding crashed caused by genomic copies.
-    (Fixes [#105](https://github.com/microPIECE-team/microPIECE/issues/101))
+    (Fixes [#105](https://github.com/microPIECE-team/microPIECE/issues/105))
 
 - [v1.2.1](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.2.1) (2018-03-23)
 
