### 2018-03-26 e156a03
Deleted fast forward seeking
