+++ scripts/CLIP_binned_bed_from_bam_and_transcripts_for_piranha.pl
@@ -193,31 +193,6 @@ sub parse_header
     return \@seq;
 }
 
-sub check4fastforward{
-    my ($start, $fast_forward, $bam, $seq, $bin_size) = @_;
-
-    while ($fast_forward>5*$binsize)
-    {
-	if ($start+$fast_forward>$seq->{len})
-	{
-	    return;
-	}
-	my $stop = $start+$fast_forward-1;
-
-	my $cmd = sprintf("samtools view -c -F 0x4 %s %s:%d-%d", $bam, $seq->{name}, $start, $stop);
-	my $counts =  qx($cmd)+0;
-	if ($? != 0)
-	{
-	    die "Error running command '$cmd'\n";
-	}
-
-	return $fast_forward if ($counts == 0);
-	$fast_forward = int($fast_forward>>1);
-    }
-
-    return;
-}
-
 sub transform2bed
 {
     my ($start, $end) = @_;
