### 2018-03-26 fc74e5d
Add parser of mirbase *.dat files to extract precursor, species, length, matures, and precursor sequence
