### 2018-03-26 #616 @luizirber
@brainstorm on it =]

@greatfireball I see there were quite a bunch of new releases since the submission, do you want me to review the latest one (`1.2.3`?)
