### 2018-03-26 #616 @luizirber
Some comments and fixes required:

# General checks

- [ ] André Billion and Andreas Vilcinskas are listed in the paper, but not in the repository. (nothing wrong with that, but maybe they should be in the repo too?)

# Installation

- [ ] Use a tagged version instead of master?
`git clone -b v1.2.3 git@github.com:microPIECE-team/microPIECE.git`
`curl -L -o microPIECE.zip https://github.com/microPIECE-team/microPIECE/archive/v1.2.3.zip`

# Functionality

- [ ] I couldn't reproduce the example in the README using v1.2.3, but it worked with v1.0.0
(both using Docker)

# Documentation

## Installation instructions

- [ ] Most of the dependencies are already in bioconda, you might consider packaging microPIECE too. (This is not a requirement for article approval, but would be nice to have).

# Software paper

## References

- [ ]  DOI are missing
