### 2018-03-27 ca7a9c2
Add a module holding the required subroutines
