### 2018-03-27 6c93153
Remove precursor copies and rename input variable
