### 2018-03-27 d470b31
Add test for MINING_complete_mirbase...
