+++ t/MINING_complete_mirbase_by_miRDeep2_output.pl
@@ -1,0 +1,11 @@
+use strict;
+use warnings;
+
+use Test::More;
+
+use FindBin;
+use lib "$FindBin::Bin/../scripts/lib";
+
+BEGIN { use_ok('mining') };
+
+done_testing;
