### 2018-03-27 27823f7
Add test for fasta importer
