### 2018-03-27 990628f
Renamed subroutines for mirbase *.dat file parsing
