### 2018-03-27 13e48fd
enhanced paper.bib with dois
