+++ paper/paper.bib
@@ -9,6 +9,8 @@
   number    = {1},
   pages     = {37--52},
   publisher = {Oxford University Press},
+  doi={10.1093/nar/gkr688},
+  url={https://doi.org/10.1093/nar/gkr688}
 }
 
 @Article{martin2011cutadapt,
@@ -19,6 +21,8 @@
   volume  = {17},
   number  = {1},
   pages   = {pp--10},
+  doi={10.14806/ej.17.1.200},
+  url={http://dx.doi.org/10.14806/ej.17.1.200}
 }
 
 @Article{pantano2009seqbuster,
@@ -30,6 +34,8 @@
   number    = {5},
   pages     = {e34--e34},
   publisher = {Oxford University Press},
+  doi={10.1093/nar/gkp1127},
+  url={https://doi.org/10.1093/nar/gkp1127}
 }
 
 @Article{kozomara2013mirbase,
@@ -41,6 +47,8 @@
   number    = {D1},
   pages     = {D68--D73},
   publisher = {Oxford University Press},
+  doi={10.1093/nar/gkt1181},
+  url={https://doi.org/10.1093/nar/gkt1181}
 }
 
 @Article{betel2008microrna,
@@ -52,6 +60,8 @@
   number    = {suppl\_1},
   pages     = {D149--D153},
   publisher = {Oxford University Press},
+  doi={10.1093/nar/gkm995},
+  url={https://doi.org/10.1093/nar/gkm995}
 }
 
 @article{wu2010fast,
@@ -62,7 +72,9 @@
   number={7},
   pages={873--881},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btq057},
+  url={https://doi.org/10.1093/bioinformatics/btq057}
 }
 
 @Article{uren2012site,
@@ -74,6 +86,8 @@
   number    = {23},
   pages     = {3013--3020},
   publisher = {Oxford University Press},
+  doi={10.1093/bioinformatics/bts569},
+  url={https://doi.org/10.1093/bioinformatics/bts569}
 }
 
 @Misc{rice2000emboss,
@@ -81,6 +95,8 @@
   title     = {EMBOSS: the European molecular biology open software suite},
   year      = {2000},
   publisher = {Elsevier Current Trends},
+  doi={10.1016/S0168-9525(00)02024-2},
+  url={https://doi.org/10.1016/S0168-9525(00)02024-2}
 }
 
 @Article{li2009fast,
@@ -92,6 +108,8 @@
   number    = {14},
   pages     = {1754--1760},
   publisher = {Oxford University Press},
+  doi={},
+  url={}
 }
 
 @Article{altschul1990basic,
@@ -103,6 +121,8 @@
   number    = {3},
   pages     = {403--410},
   publisher = {Elsevier},
+  doi={10.1016/S0022-2836(05)80360-2},
+  url={https://doi.org/10.1016/S0022-2836(05)80360-2}
 }
 
 @Article{li2009sequence,
@@ -114,6 +134,8 @@
   number    = {16},
   pages     = {2078--2079},
   publisher = {Oxford University Press},
+  doi={10.1093/bioinformatics/btp352},
+  url={https://doi.org/10.1093/bioinformatics/btp352}
 }
 
 @Article{quinlan2010bedtools,
@@ -125,6 +147,8 @@
   number    = {6},
   pages     = {841--842},
   publisher = {Oxford University Press},
+  doi={10.1093/bioinformatics/btq033},
+  url={https://doi.org/10.1093/bioinformatics/btq033}
 }
 
 @Article{lechner2011proteinortho,
@@ -136,6 +160,8 @@
   number    = {1},
   pages     = {124},
[…]
