### 2018-03-27 #616 @DanielAmsel
Thank you @luizirber for your review. We are going to fix the raised issues as soon as possible.
