### 2018-03-27 f63ebdf
Refactoring of mirDeep2 parser to use a central version
