+++ scripts/MINING_curate_mirdeep2fasta.pl
@@ -4,6 +4,11 @@ use warnings;
 use Getopt::Long;
 use Digest::MD5;
 
+use FindBin;
+use lib "$FindBin::Bin/lib";
+
+use mining;
+
 my $csv_file;
 my $cutoff;
 my $mature_file;
@@ -28,7 +33,7 @@ die "Hairpin file exists and will not be overwritten!\n" if (-e $hairpin_file);
 die "Need to specify --species three-letter-species-code\n" unless (defined $species);
 
 # get novel miRNAs above threshold
-my $novels = parse_mirdeep($csv_file, $cutoff);
+my $novels = mining::parse_mirdeep_novels($csv_file, $cutoff);
 
 open(MATURE,">",$mature_file)   || die "Unable to open file '$mature_file': $!\n";
 open(HAIRPIN,">",$hairpin_file) || die "Unable to open file '$hairpin_file': $!\n";
@@ -52,102 +57,3 @@ for(my $novel_count = 1; $novel_count <= @{$novels}; $novel_count++)
 }
 close(MATURE) || die "Unable to close file '$mature_file': $!\n";
 close(HAIRPIN)|| die "Unable to close file '$hairpin_file': $!\n";
-
-###########################################################################
-
-# skip all parts but novel section
-sub parse_mirdeep{
-        my ($file, $cutoff)     = @_;
-        my @result = ();
-	my %seen   = ();
-
-        open(PM, "<", $file) || die "Unable to open file '$file': $!\n";
-        while(<PM>){
-	    next unless (/^provisional/);
-
-	    while(<PM>)
-	    {
-		chomp;
-
-		# leave the inner loop if the line is empty
-		last if (/^\s*$/);
-
-		my %dataset = ();
-
-		my @fieldnames = (
-		    "provisional_id",                         # provisional id
-		    "score",                                  # miRDeep2 score
-		    "probability",                            # estimated probability that the miRNA candidate is a true positive
-		    "rfam_alert",                             # rfam alert
-		    "total_read_count",                       # total read count
-		    "mature_count",                           # mature read count
-		    "loop_count",                             # loop read count
-		    "star_count",                             # star read count
-		    "randfold_significant",                   # significant randfold p-value
-		    "mirbase_mirna",                          # miRBase miRNA
-		    "mirbase_example",                        # example miRBase miRNA with the same seed
-		    "UCSC_browser",                           # UCSC browser
-		    "NCBI_blastn",                            # NCBI blastn
-		    "mature_seq",                             # consensus mature sequence
-		    "star_seq",                               # consensus star sequence
-		    "precursor_seq",                          # consensus precursor sequence
-		    "precursor_coordinates"                   # precursor coordinate
-		    );
-
-
-		@dataset{@fieldnames} = split("\t", $_);
-
-		next if ($dataset{score} < $cutoff);
-
-		# check if we need to switch mature and star
-		my $mature_idx	= index($dataset{precursor_seq},$dataset{mature_seq});
-		my $star_idx	= index($dataset{precursor_seq},$dataset{star_seq});
-
-		if ($star_idx == $mature_idx)
-		{
-		    die("mature and star sequence have the same position on hairpin\n"); # should never happen
-		}
-		# we need to switch if star is before mature
-		if($star_idx < $mature_idx){
-		    warn("Need to switch mature/star sequence for line '$_'\n");
-		    ($dataset{mature_seq}, $dataset{star_seq}) = ($dataset{star_seq}, $dataset{mature_seq});
-		}
-
-		# calculate a checksum for "hairpin_seq|mature_seq|star_seq"
-		my $ctx = Digest::MD5->new;
[…]
