### 2018-03-27 4dd27e8
Remove switching of mature/star sequence based on position from mirdeep2-parser
