### 2018-03-27 91e842b
Refactoring of the mirbase known miRNAs handling
New matures are now only attached, if both matures have

1) 10 read counts each
or
2) a sum of 100 read counts and each a minimum read count of 5

This decission is based on mirBase' definition of the high confidence set:
http://www.mirbase.org/blog/2014/07/high-confidence-mirna-set-available-for-mirbase-21/
