+++ scripts/MINING_complete_mirbase_by_miRDeep2_output.pl
@@ -19,126 +19,99 @@ GetOptions(
 
 my $mirbase_dat_content = mining::parse_mirbase_dat($mirbase_dat, $species);
 
-__END__
-
-# split the list of given precursors that have genomic copies
-my @precursor_list	= split(",",$precursor_copies);
-my %precursor_hash;	#{precursor} = 1
-foreach my $precursor (@precursor_list){
-	$precursor =~s/mir/miR/;
-	$precursor =~s/-\d$//;
-	$precursor_hash{$precursor} += 1;
+my %mature2precursorIndex = ();
+for(my $i=0; $i<@{$mirbase_dat_content}; $i++)
+{
+    foreach my $mature (@{$mirbase_dat_content->[$i]{matures}})
+    {
+	push(@{$mature2precursorIndex{$mature->{name}}}, $i);
+    }
 }
 
-my %mature_hash		= %{&parse_fasta($mature_fasta)};
-my %pair_hash		= %{&get_pairs(\%mature_hash)};
+my $mirdeep_content = mining::parse_mirdeep_known($mirdeep_csv);
 
-open(CSV,"<",$mirdeep_csv) || die;
-my $line_bool = 0;
-while(<CSV>){
-	chomp;
-	if(/^mature miRBase miRNAs detected by miRDeep2/){
-		$line_bool = 1;
-		next;
+foreach my $mirdeep_line (@{$mirdeep_content})
+{
+    unless (exists $mature2precursorIndex{$mirdeep_line->{mirbase_mirna}})
+    {
+	die "Unable to find an entry for $mirdeep_line->{mirbase_mirna}\n";
+    }
+
+    # get a list of potential precursor
+    my @precursor = @{$mature2precursorIndex{$mirdeep_line->{mirbase_mirna}}};
+
+    # filter for precursors without two mature sequence
+    my @precursor_2_complete = grep { int(@{$mirbase_dat_content->[$_]{matures}}) < 2 } (@precursor);
+
+    # for the precursors_2_complete, we want add another mature from mirdeep
+    if (@precursor_2_complete)
+    {
+
+	# check for a minimum of 10 reads each for mature/star sequence or sum of both > 0 and both minimum 5
+	if (
+	    ($mirdeep_line->{mature_count} >= 10 && $mirdeep_line->{star_count} >= 10)
+	    ||
+	    ($mirdeep_line->{mature_count}+$mirdeep_line->{star_count}>=100 && $mirdeep_line->{mature_count}>=5 && $mirdeep_line->{star_count}>=5)
+	    )
+	{
+	    warn(sprintf("Mature/star sequences for '%s' do have enough counts (%d/%d) to fulfil requirement for adding mature sequence\n", $mirdeep_line->{mirbase_mirna}, $mirdeep_line->{mature_count}, $mirdeep_line->{star_count}));
+	} else {
+	    warn(sprintf("Mature/star sequences for '%s' do not have enough counts (%d/%d) to fulfil requirement for adding mature sequence\n", $mirdeep_line->{mirbase_mirna}, $mirdeep_line->{mature_count}, $mirdeep_line->{star_count}));
+	    next;
 	}
-	next if ($line_bool ==0);
-	next if (/^tag/);
-	if(/^$/){
-		$line_bool = 0;
-		next;
+
+	warn("Adding missing mature sequences for mirDeep2 mature ".$mirdeep_line->{mirbase_mirna}."\n");
+
+	if (@precursor > 1 && @precursor_2_complete != @precursor)
+	{
+	    warn("Multiple potential precursors with some not having 2 matures for '".$mirdeep_line->{mirbase_mirna}."\n");
[…]
