+++ scripts/MINING_complete_mirbase_by_miRDeep2_output.pl
@@ -118,3 +118,8 @@ foreach my $mirdeep_line (@{$mirdeep_content})
 	}
     }
 }
+
+my $output = "";
+mining::export_mirbase_dat(\$output, $mirbase_dat_content);
+
+print $output;
