### 2018-03-28 6bf9252
Fix division by zero bug during quantification
Fix #118

This fix avoids the division by zero exception if libraries can not be mapped on the reference
