### 2018-03-28 9ffdabb
Fix typo for --piranhabinsize parameter
Fix #116
