+++ microPIECE.pl
@@ -74,7 +74,7 @@ GetOptions(
     'speciesBtag=s'        => \$opt->{speciesB_tag},
     'mirbasedir=s'         => \$opt->{mirbasedir},
     'tempdir=s'            => \$opt->{tempdir},
-    'piranahbinsize=i'     => \$opt->{piranha_bin_size},
+    'piranhabinsize=i'     => \$opt->{piranha_bin_size},
     ) || pod2usage(1);
 
 # split clip files if required
@@ -257,9 +257,9 @@ The folder specified by C<--tempdir> is used for temporary files. The
 default value is F<tmp/> inside the output folder specified by the
 C<--out> parameter.
 
-=item C<--piranahbinsize>
+=item C<--piranhabinsize>
 
-Sets the F<Piranah> bin size and has a default value of C<20>.
+Sets the F<Piranha> bin size and has a default value of C<20>.
 
 =back
 
@@ -448,7 +448,7 @@ L<#34|https://github.com/microPIECE-team/microPIECE/issues/34>)
 
 =item L<v1.0.7|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.7> (2018-03-08)
 
-Piranha was lacking of a bin_size parameter. Added parameter C<--piranahbinsize> with a default value of C<20>
+Piranha was lacking of a bin_size parameter. Added parameter C<--piranhabinsize> with a default value of C<20>
 (Fixes L<#66|https://github.com/microPIECE-team/microPIECE/issues/80>)
 
 =item L<v1.0.6|https://github.com/microPIECE-team/microPIECE/releases/tag/v1.0.6> (2018-03-08)
