### 2018-03-28 microPIECE-team/microPIECE#123 @DanielAmsel
change docker version from master to tagged
The example needs a tagged version for testing, instead of master.

https://github.com/openjournals/joss-reviews/issues/616
