### 2018-03-28 microPIECE-team/microPIECE#123 @DanielAmsel
considered in pull requests
#121 and #122
