### 2018-03-28 c1d5afa
Implemented datout parameter
