### 2018-03-28 c9e44e6
Add writing of mirbase-pseudo-dat for novels
