+++ scripts/MINING_curate_mirdeep2fasta.pl
@@ -62,9 +62,35 @@ for(my $novel_count = 1; $novel_count <= @{$novels}; $novel_count++)
 
     my $header	= sprintf("%s-new-%s", $species, $novel->{digest});
 
+    my %dataset = (
+	precursor => $header,
+	seq       => $hairpin,
+	species   => uc($species),
+	matures   => [
+	    {
+		name => $header."-5p",
+		start => index($hairpin, $mature5p)+1,                       # zero based from index, but need to be one-bases
+		stop  => index($hairpin, $mature5p)+length($mature5p)        # start+length-1
+	    },
+	    {
+		name => $header."-3p",
+		start => index($hairpin, $mature3p)+1,                       # zero based from index, but need to be one-bases
+		stop  => index($hairpin, $mature3p)+length($mature3p)        # start+length-1
+	    }
+	]
+	);
+    push(@{$dat_info}, \%dataset);
+
     print HAIRPIN ">", $header, "\n", $hairpin, "\n";
     print MATURE  ">", $header, "-5p\n", $mature5p, "\n";
     print MATURE  ">", $header, "-3p\n", $mature3p, "\n";
 }
 close(MATURE) || die "Unable to close file '$mature_file': $!\n";
 close(HAIRPIN)|| die "Unable to close file '$hairpin_file': $!\n";
+
+my $dat_tmp = "";
+mining::export_mirbase_data(\$dat_tmp, $dat_info);
+
+open(FH, ">>", $datout) || die "Unable to open file '$datout' for appending: $!\n";
+print FH $dat_tmp;
+close(FH) || die "Unable to close file '$datout' for appending: $!\n";
