### 2018-03-28 microPIECE-team/microPIECE#128 @DanielAmsel
DOI are missing
paper.bib needs DOI

Issue raised in #https://github.com/openjournals/joss-reviews/issues/616

pull-request #115
