### 2018-03-28 327247a
Refactory of the ISOMIR_create_mirbase_struct.pl script
Input data are now pseudo-mirbase-dat files
