+++ scripts/ISOMIR_create_mirbase_struct.pl
@@ -1,17 +1,22 @@
-#! /usr/bin/perl
+#!/usr/bin/env perl
 use strict;
 use warnings;
 use RNA::HairpinFigure qw/draw/;
 use Getopt::Long;
 use File::Temp qw(tmpnam);
 
+use FindBin;
+use lib "$FindBin::Bin/lib";
+
+use mining;
+
 #>cel-let-7 (-42.90)   [cel-let-7-5p:17-38] [cel-let-7-3p:60-81]
 #
-#------uaca    gga             U              ---  aaua 
+#------uaca    gga             U              ---  aaua
 #          cugu   uccggUGAGGUAG AGGUUGUAUAGUUu   gg    u
-#          ||||   ||||||||||||| ||||||||||||||   ||     
+#          ||||   ||||||||||||| ||||||||||||||   ||
 #          gaca   aggCCAUUCCAUC UUUAACGUAUCaag   cc    u
-#agcuucucaa    --g             U              ugg  acca 
+#agcuucucaa    --g             U              ugg  acca
 
 
 #my $name   = 'hsa-mir-92a-1 MI0000093 Homo sapiens miR-92a-1 stem-loop';
@@ -25,148 +30,53 @@ use File::Temp qw(tmpnam);
 #GGUUGGCAUAAGGUGGUACCAUGUAACAUUUUAACCCAUAGUACGACCCAUGCCGACUCA
 #(((((((((..(((.((((.(((.............))).)))).))).))))))))).. (-21.92)
 
-my $hairpin_file;
-my $mature_file;
-my $struct_file;
-my $custom_tmp;
+my $output;
+my $mirbase_dat;
 
 GetOptions(
-    "hairpin=s"	=> \$hairpin_file,
-    "mature=s"	=> \$mature_file,
-    "struct=s"	=> \$struct_file,
-    "out=s"         => \$custom_tmp
+    "mirbasedat=s" => \$mirbase_dat,
+    "out=s"        => \$output
     ) || die;
 
-die "Need to specify a existing hairpin file via --hairpin parameter\n" unless (defined $hairpin_file && -e $hairpin_file);
-die "Need to specify a existing mature file via --mature parameter\n" unless (defined $mature_file && -e $mature_file);
-die "Need to specify a existing structure file via --struct parameter\n" unless (defined $struct_file && -e $struct_file);
-die "Need to specify a non-existing output file via --out parameter\n" unless (defined $custom_tmp && ( ! -e $custom_tmp ));
+die "Need to specify a existing mirbase file via --mirbasedat parameter\n" unless (defined $mirbase_dat && -e $mirbase_dat);
+die "Need to specify a non-existing output file via --out parameter\n" unless (defined $output && ( ! -e $output ));
 
 my $rnafold		= "RNAfold -noPS";
 
+my $mirnas = mining::parse_mirbase_dat($mirbase_dat, undef);
 
-my %hairpin_hash	= %{&read_fasta($hairpin_file)};
-my %mature_hash		= %{&read_fasta($mature_file)};
+my $foldinput = tmpnam();
 
-my %structure_hash	= %{&parse_structure($struct_file)};
+open(OUT,">",$output) || die "Unable to open file '$output' for writing: $!\n";
 
-open(OUT,">",$custom_tmp) || die;
+foreach my $entry (@{$mirnas})
+{
+    open(FH, ">", $foldinput) || die "Unable to open file '$foldinput': $!\n";
+    print FH ">temp\n", $entry->{seq}, "\n";
[…]
