+++ microPIECE.pl
@@ -417,6 +417,8 @@ Please report any new issues ad L<new Github-Issue|https://github.com/microPIECE
 
 =item scheduled for next release
 
+Creating all structures on the fly using pseudo-mirBASE-dat as input.
+
 Using F<miRNA.dat> from mirBASE as source for mature/precursor sequence and relationship (Fixes L<#127|https://github.com/microPIECE-team/microPIECE/issues/127>)
 
 Fix of division-by-zero bug for empty mapping files (Fixes L<#118|https://github.com/microPIECE-team/microPIECE/issues/118>)
