### 2018-03-28 ca4d3cb
Preparation for release v1.3.0
