### 2018-03-29 cf723c0
Estimate ProgressBar width to handle redirect error
