### 2018-03-29 b0f357a
Introduced a silient/quiet parameter
