### 2018-03-31 cfeb3e9
Bump docker image to v1.4.0
