### 2018-04-05 e949651
Update paper.bib
