+++ paper/paper.bib
@@ -175,5 +175,17 @@
   doi={10.1073/pnas.1701474114},
   url={https://doi.org/10.1073/pnas.1701474114}
 }
+@article{amsel2017evaluation,
+  title={Evaluation of high-throughput isomiR identification tools: illuminating the early isomiRome of Tribolium castaneum},
+  author={Amsel, Daniel and Vilcinskas, Andreas and Billion, Andr{\'e}},
+  journal={BMC bioinformatics},
+  volume={18},
+  number={1},
+  pages={359},
+  year={2017},
+  publisher={BioMed Central},
+  doi={10.1186/s12859-017-1772-z},
+  url={https://doi.org/10.1186/s12859-017-1772-z}
+}
 
 @Comment{jabref-meta: databaseType:bibtex;}
