+++ paper/paper.md
@@ -41,7 +41,7 @@ In case of a provided smallRNA dataset, the pipeline starts with the miRNA analy
 
 If no smallRNA dataset is provided, the pipeline directly jumps to the CLIP analysis. There it starts with the *speciesA* CLIP-seq library trimming, using `Cutadapt` [@martin2011cutadapt]. Trimmed reads are then mapped to the genome with `gsnap` [@wu2010fast] and the results are evaluated by `Piranha` [@uren2012site]. Then the libraries are merged into the `BED` file format. We further used `SAMtools` [@li2009sequence] and `BEDtools` [@quinlan2010bedtools] for file conversions during the pipeline. The `BED` file includes a column that displays how many libraries support each genomic position. Next, a file for each library-support-level is created, so that the user can in the end decide how many CLIP libraries are necessary to account this region as binding region. Now for each library-support-level, an assignment of each sequence to the transcriptome is performed. Outgoing from the transcript, the corresponding protein is used to discover the orthologous protein in the *speciesB* by `Proteinortho` [@lechner2011proteinortho]. This information is used as criteria to align the CLIP region from *speciesA* to the orthologous transcript in *speciesB* with `EMBOSS Needle` [@rice2000emboss].
 
-Finally, the pipeline passes the miRNA set (either from the *full* or *minimal workflow*) to `miraligner` from the `seqbuster` package [@pantano2009seqbuster]. Then a target prediction with `miranda` [@betel2008microrna] on the previously transferred orthologous CLIP regions is performed. 
+The pipeline analizes the miRNA set (either from the *full* or *minimal workflow*) for miRNA isoforms. Outgoing from previous benchmarks [@amsel2017evaluation], we chose `miraligner` from the `seqbuster` package [@pantano2009seqbuster] as miRNA isoform detection tool. Finally, a target prediction with `miranda` [@betel2008microrna] on the previously transferred orthologous CLIP regions is performed. 
 
 Depending on the provided data, the minimal output of the pipeline consists of a target prediction output from `miranda` for each library-support-level, based on the transferred CLIP regions. In case the pipeline additionally received smallRNA data, a microRNA set with known and novel miRNAs together with an expression file is saved, as well as the orthologs to other species, the genomic loci of the miRNAs and the identified isoforms.
 
