### 2018-04-05 36fe7ac
Move CLIP_bedtool_discrard_sizes.pl into microPIECE module
