### 2018-04-05 747dcfe
Update paper.md
