### 2018-04-05 #616 @greatfireball
Thanks @luizirber for your review. I think we addressed most of your issues, but we need some help with your issue concerning the example (please see below about our detailed answers). In between, we further improved `microPIECE` and pushed our current version to `v1.4.0`. Moreover, we added an explaination why we chose `miraligner` as isomer detection tool into the manuscript.

## General checks

- [X]  André Billion and Andreas Vilcinskas are listed in the paper, but not in the repository. (nothing wrong with that, but maybe they should be in the repo too?)

**Answer**: *Neither André nor Andreas wrote a single line of code, but they have been involved in developing the method. Therefore, they own not a single commit or comment in our repository, but are co-authors of our paper*

## Installation

- [X] Use a tagged version instead of master?
```
git clone -b v1.2.3 git@github.com:microPIECE-team/microPIECE.git
curl -L -o microPIECE.zip https://github.com/microPIECE-team/microPIECE/archive/v1.2.3.zip
```
**Answer**: *We updated the installation instructions and used the current tag. This modification includes the usage of tags for the docker containers.*

## Functionality
- [ ] I couldn't reproduce the example in the README using v1.2.3, but it worked with v1.0.0
    (both using Docker)

**Answer**: *To be honest, we tested the example using different operation systems (Ubuntu/OSX) on different machines and had not a single failure running our example. Can you give a little more details about your setup and what step failed?*

## Documentation
### Installation instructions

- [ ] Most of the dependencies are already in bioconda, you might consider packaging microPIECE too. (This is not a requirement for article approval, but would be nice to have).

**Answer**: *I will check for all dependencies and will create a bioconda recipe in the near future. If finished, we will include an installation via bioconda as alternative into our documentation*

## Software paper
### References
- [X] DOI are missing

**Answer**: *We included the missing DOIs*
