### 2018-04-05 7836e24
Add a test to ensure valid parameters for --CLIPminProcessLength and --CLIPmaxProcessLength
