### 2018-04-05 d3a260a
Implement usage of user provided limit for run_CLIP_clip_mapper
