### 2018-04-05 a80dd7a
Update documentation
Fix #145
