### 2018-04-06 5f4fab3
Added a pre-sort step for each bin and keep it in memory
