### 2018-04-09 #616 @luizirber
Thanks for the comments and updates @greatfireball and @DanielAmsel!

I reran the docker example and it worked now, may have been a glitch on my side (I don't have the error message anymore...)

@brainstorm I checked all the boxes, I think this is good to go! (can you update the version in the initial comment of this PR? to `1.4.0`?)
