### 2018-04-10 #616 @greatfireball
Good morning! Thanks @luizirber and @brainstorm.

If possible, I would like to release and use the version v1.5.0, due to we further optimized our code. I will generate a new release later this morning, create a new zenodo archive, and post a link and the DOI here.

Thanks!
