### 2018-04-10 #616 @brainstorm
Good afternoon from Melbourne!

Thanks @greatfireball, apparently the Zenodo DOI is stable across versions and points to the latest version:

```
Cite all versions? You can cite all versions by using the DOI 10.5281/zenodo.1188480. This DOI represents all versions, and will always resolve to the latest one.
```

As I just read here: http://help.zenodo.org/#versioning

@arfon, does whedon announce the release of the paper upon closing this issue I assume, or is there any additional step to follow?

Thanks everyone again!
