### 2018-04-10 #616 @greatfireball
@brainstorm of course you are right... You set the stabil DOI. I thought you would like to use a DOI pointing to the version indicated by the version above. Do you increase the version number to v1.5.0 before closing the issue?
