### 2018-04-10 506e156
Refactored CLIP_merge_bed_files.pl
The huge memory consumption should be solved, by presorting and working on just a single chromosome.

Fixes #174
