+++ scripts/CLIP_merge_bed_files.pl
@@ -137,10 +137,9 @@ if ($output ne "-")
     }
 }
 
-my @genome = ();
 my %chromosomes = ();
+my @genome = ();
 my %strands    = ();
-my %conditions = ();
 
 my %input_sorted = ();
 my $max_length = 0;
@@ -163,103 +162,91 @@ foreach my $key (keys %input)
 	{
 	    $chromosomes{$set->[0]} = $set->[2] unless (exists $chromosomes{$set->[0]} && $chromosomes{$set->[0]} >= $set->[2]);
 	    $max_length = $set->[2] if ($max_length < $set->[2]);
+
+	    # store strands
+	    $strands{$set->[5]}++;
 	}
 
-	push(@{$input_sorted{$key}} , join("", map { join("\t", @{$_}) } (@temp)));
+	push(@{$input_sorted{$key}{dat}} , \@temp);
+	push(@{$input_sorted{$key}{pos}} , 0);
     }
 }
 
 my @chromosome_order = sort {$a cmp $b} (keys %chromosomes);
 WARN(sprintf("Found a total of %d chromosomes with a maximum length (covered) of %d bp. Expected chromosome order: %s", int(@chromosome_order), $max_length, join(", ", map {"'$_'"} (@chromosome_order))));
 
-foreach my $key (keys %input_sorted)
-{
-    foreach my $file (@{$input_sorted{$key}})
-    {
-	open(FH, "<", \$file) || LOGDIE("Unable to open file FILEHANDLE: $!");
-	while(<FH>)
-	{
-	    # go through the bed files
-	    chomp;
+# map condition to number
+my @conditions_ordered = sort (keys %input_sorted);
+my %condition2number;
+@condition2number{@conditions_ordered} = (0..(int(@conditions_ordered)-1));
 
-	    # NW_001809801.1	79510	79534	XM_001647792.1	1	+
-	    # NW_001809801.1	79527	79534	XM_001647792.1	1	+
-	    # NW_001809801.1	248783	248807	XM_001647796.1	1	-
-            # NW_001809801.1	533942	533950	XM_001647802.1	1	-
-            # NW_001809801.1	2779750	2779790	XM_001647839.1	1	+
+# map strand to number
+my @strands_ordered = sort (keys %strands);
+my %strand2number;
+@strand2number{@strands_ordered} = (0..(int(@strands_ordered)-1));
 
-	    my ($chromosome, $start, $stop, $mrna_ids, undef, $strand) = split(/\t/, $_);
-
-	    # check if chromosome is already known
-	    unless (exists $chromosomes{$chromosome})
-	    {
-		$chromosomes{$chromosome} = int(keys %chromosomes);
-	    }
-	    $chromosome = $chromosomes{$chromosome};
-
-	    # check if strand is already known
-	    unless (exists $strands{$strand})
-	    {
-		$strands{$strand} = int(keys %strands);
-	    }
-	    $strand = $strands{$strand};
-
-	    # check if condition/key is already known
-	    unless (exists $conditions{$key})
-	    {
-		$conditions{$key} = int(keys %conditions);
-	    }
-	    my $condition = $conditions{$key};
-
-	    # for each chromosomal position we need to consider:
-	    # Counter + strand
-	    # Counter - strand
-
-	    # due to bed stop field is 0-based, but exclusive, we
-	    # need to continue as long as $i is less then stop-field
-	    for (my $i=$start; $i<$stop; $i++)
-	    {
-		$genome[$chromosome][$strand][$i][$condition]++;
-	    }
-	}
-	close(FH) || LOGDIE("Unable to close FILEHANDLE: $!");
-    }
-}
-
-# print the output
-my @conditions_ordered = sort (keys %conditions);
[…]
