### 2018-04-10 b47763e
Dockerfile information bumped to v1.5.0
