### 2018-04-10 c04cf2f
Update Docker information in README.md
