### 2018-04-10 5ed03e7
Update example to use docker container for v1.5.0
