### 2018-04-10 feca1e1
Remove old prebinning script
