### 2018-04-10 bf59a00
Update documentation of CLIP_merge_bed_files.pl
This was done to reflect renaming of the old script

Fix #156
