### 2018-04-10 a4a9f2c
Add version number and help parameter
