### 2018-04-10 6eb5469
Add test for speciefied BAM files
