### 2018-04-10 b6fbd50
Add automatic determination of Docker container version
