### 2018-04-10 #616 @greatfireball
I just released the version [v1.5.0](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.5.0). The DOI for that version is [10.5281/zenodo.1215850](https://doi.org/10.5281/zenodo.1215850) but if you are fine with the DOI for all versions, we are fine as well. Finally, we will run the pipeline once more and update the numbers within the manuscript. Afterwards, we are ready to go. Thanks!
