### 2018-04-10 #616 @arfon
> @arfon, does whedon announce the release of the paper upon closing this issue I assume, or is there any additional step to follow?

@brainstorm - I have to do a couple of things in the background. I can take it from here...
