### 2018-04-10 #616 @arfon
> Thank you for finalizing the paper and thanks for acception @arfon! Nevertheless, as I stated above, we are currently rerunning the pipeline with version v1.5.0 and might need to change a number inside the manuscript. Is it still possible or to late? Thanks in advance!

Sure thing. I can update the paper whenever you're ready.
