### 2018-04-15 #616 @greatfireball
Good morning @arfon,

finally we finished the rerun and updated the numbers. Unfortunately, I forgot to merge an important pull request which dramatically reduced the runtime of a single step. Therefore, I had to create another release [v1.5.2](https://github.com/microPIECE-team/microPIECE/releases/tag/v1.5.2) and a zenodo upload [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1218731.svg)](https://doi.org/10.5281/zenodo.1218731).
If you are able to update the DOI again, I would suggest to set it to the new one: [10.5281/zenodo.1218731](https://doi.org/10.5281/zenodo.1218731) and update the version of the pipeline to v1.5.2. The manuscript is up to date now, and we also included the ORCID for André.

Thanks again to @brainstorm, @luizirber, and @arfon !

Best,
Frank
